requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
black==23.1.0
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
idna==3.4
joblib==1.1.0
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.4
mypy==0.971
mypy-extensions==1.0.0
numpy==1.23.5
packaging==23.0
pandas==1.5.2
pathspec==0.11.1
platformdirs==3.1.1
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.12
yarl==1.8.2
